requirements.txt
python-dotenv
pyTelegramBotAPI
glaider
pymailtm
requests